Changes to be committed: 	new file:   LOAD.py 	modified:   UsbControl/LightCrafter6500Device.py Made a load function to load pictures on the fly.
I fixed a bug in LightCrafter6500Device that made the DMD unable to display multiple images

I added two features to LightCrafter as well: the first feature makes the user able to reset the board.
The second feature programs the TRIGGER_OUT_1 with +50us on delay and -20us off delay.
This delay is needed to make the DMD work nicely with the Clock Thief Device